--- src/pages/QuickGift/QuickGiftPage.js
@@ -1,9 +1,0 @@
-import QuickGift from './QuickGift';
-
-export default function QuickGiftPage() {
-  return (
-    <div>
-      <QuickGift />
-    </div>
-  );
-}
